requirements.txt
uvicorn
pydantic-settings
python-dotenv
pymongo
motor
groq
openai
langchain
langchain-groq
langchain-openai
pydantic
langchain-community
langchain-mongodb
pypdf
tiktoken
sentence-transformers
langchain-huggingface
langgraph
langfuse
trulens
streamlit
